Delay registering commerce_customer addressfield
Fixes #42

This is an issue with the timing of things! If we haven't already
enabled addressfield when we enable commerce_customer, Backdrop tries to
register field_commerce_customer_address.json from the config directory
before the addressfield module has made the addressfield field type
available. This results in an error that the addressfield module isn't
available (which it isn't!).

We need to delay the loading of the field config file until all modules
are fully installed and field types are registered, and we can do that
by moving it to a hook_update().

We need to move that config file into a subdirectory of config, which
I'm not crazy about, but this works.