Fixes for Commerce Customer submodule config
Our original port incorrectly referenced config in Commerce Checkout
that didn't exist. This commit adds this config to the Commerce Customer
submodule where it belongs, and updates the code to point to the correct
configuration file.

This will fix an issue where Commerce orders and addresses were getting
stored in the database, but weren't linked to one another. It also
should solve the profile copying issues that we detected during a failed
test that is referenced in Issue 29.

We know a few users have the unreleased version of this module, so we're
adding an update hook to commerce_customer to make sure they receive
this change as well on upgrade.