Fix for Commerce token population
This fixes #52 where tokens weren't populating on Orders.

hook_tokens() takes a $data array of objects, and the code checks to see
if $type is present in the array.

In Drupal 7, the $data array often had multiple keys. The ones supplied
by the Rules module, for example, have dashed keys, e.g. my-entity. The
ones supplied by the Salesforce Suite module had underscores, e.g.
my_entity. The Drupal 7 code worked because both keys were available in
the $data array.

In Backdrop, $data only contains one array, and we were encountering
mismatches of dashes and underscores. As a result, Commerce code was
exiting early instead of replacing our tokens with the expected values,
because it wasn't finding the type it was expecting in the $data array
to begin with.

This commit updates Commerce to check for both dashed and underscored
versions of the type when looking for the correct entity type in $data.

It also fixes a bug with chained tokens: first, a strucutral fix so
that the switch code will run, and second, an issue with an array union:
the module was setting the default owner field value to "name", and even
when that value got set via token_replace(), the array union (+=) was
clobbering the new value with the default name value.